0.8.9

+++ app/src/main/java/com/hazizz/droid/Communication/POJO/Response/POJOuser.java
@@ -4,12 +4,12 @@ import lombok.Data;
 
 @Data
 public class POJOuser {
-    private int id;
+    private long id;
     private String username;
     private String displayName;
     private String registrationDate;
 
-    POJOuser(int id, String username, String registrationDate){
+    POJOuser(long id, String username, String registrationDate){
         this.id = id;
         this.username = username;
         this.registrationDate = registrationDate;
